feat: implement complete Compose UI with full pipeline integration
Implement three-screen bottom navigation architecture with integrated
Google Maps visualization and full detection pipeline.

Screens:
- ConfigScreen: Route file management, parameter sliders with live updates
- DetectionScreen: 60/40 map/status split with real-time position
- HistoryScreen: Time-filtered event log with day grouping

Data Layer:
- RouteStorageManager: Route file copy, metadata CRUD operations
- DetectionPreferences: Parameter persistence via SharedPreferences
- DetectionRepository: Room database for arrival/departure history

ViewModels:
- ConfigViewModel: Route CRUD, active route selection, parameter updates
- DetectionViewModel: Service binding, active route loading, event stream
- HistoryViewModel: Time-based filtering (Today/Week/All)

Service Integration:
- DetectionService: Full pipeline (MapMatcher → Kalman → StateMachine)
- GeoCoordinateConverter: Lat/lon to grid projection
- Service binding for real-time event emission

UI Components:
- MapView: Google Maps with route polyline, stop markers, position
- StatusPanel: Service status, position display, event stream
- RouteListItem, ParameterSlider, FilterChips, EventItem

Testing:
- Unit tests for all ViewModels (Config, Detection, History)
- UI tests for all screens (Compose testing)
- Turbine for Flow testing

Polish:
- Loading states with spinners and skeleton screens
- Empty states with illustrations and helpful text
- Error handling with snackbar and dialog improvements
- Animations (fade-in, expand/collapse, staggered lists)

Production:
- Release signing configuration with keystore.properties
- ProGuard rules for Compose, Room, Hilt, Gson, Maps
- Build variants (debug/release) with proper suffixes
- Production build documentation

Pipeline flow: Location → GeoConverter → MapMatcher → Kalman → StateMachine → Events

+++ android/app/src/main/java/com/busarrival/app/domain/model/GpsPoint.kt
@@ -32,6 +32,6 @@ data class GpsPoint(
 }
 
 // Extension functions for Location conversion
-private fun Double.toCdeg(): Short = (this * 100).toShort()
-private fun Float.toCdeg(): Short? = if (this >= 0) (this * 100).toShort() else null
+private fun Double.toCdeg(): Short = (this * 100).toInt().toShort()
+private fun Float.toCdeg(): Short? = if (this >= 0) (this * 100).toInt().toShort() else null
 private fun Float.toCms(): Int? = if (this >= 0) (this * 100).toInt() else null
